Split condition validation
